requirements.txt
Django==1.5.1
PyJWT==0.1.5
South==0.8.1
dj-database-url==0.2.1
gevent==0.13.8
psycopg2==2.5.1
python-dateutil==2.1
pytz==2013b
requests==1.2.3
shortuuid==0.3
six==1.3.0
gunicorn==0.17.4
honcho==0.4.2
newrelic==1.12.0.56
boto==2.9.6
django-storages==1.1.8
django-admin-bootstrapped==0.4.1
hiredis==0.1.1
redis==2.7.6
django-redis-sessions==0.4.0
